Major change to add additional bonus output datatypes that can be provided by any datatype, stored in the same directory. Makes it easy for a datatype to provide several different views of its data suitable for input to different modules, or.
Added simple copy file module to manage output files.
First example of an additional datatype provided by the plot module, which supplied one for each of its individual files, which can then be referenced in a config and used as input (e.g. to the copy_file module).

Fixed a problem with getting output from multistage modules which meant it was always instantiated with the wrong base dir.