Major changes to parameter alternatives
Enhancements of parameter alternatives in config files. Added option to tie alternatives. Added naming of alternative values, used to put together expanded module name.

Added implicit alternative inputs where an input module has been expanded for alternative params.

Completely restructured loading of pipeline config, so that module infos are instantiated during pipeline loading and not when the PipelineConfig is instantiated. This fits much better with the way pipelines are now used - the old system was a remnant from a previous idea of how modules would be loaded.

Note that the rewriting of the pipeline loading code will probably affect other things that I've not taken account of, so the codebase may be unstable for a while.

Changed expanded module naming to use []s insteadof {}, so we don't need to worry about the {}s being misinterpreted in string formatting.

Added a feature to allow you to specify modules by their numbers in the CLI, instead of typing the full name, which is a pain.