Added all-deps swtich and dry-run
Added a switch to the run command to add to the list of modules to be executed any from earlier in the pipeline that are not currently executed, but that the given modules depend on.

Also deprecated the check command in favour of a --dry-run switch to the run command, so that the checks performed are guaranteed identical to what happens before running. Also means it can be combined with e.g. --all-deps.