New module variables feature
Added module reference to datatypes, so that it's easy to trace which module they were created as output to (if any).

Added new feature to pipeline config syntax: module variables. Using some special parameter types (starting 'modvar_'), you can now associate a dictionary of variable assignments with a module. These get passed through the pipeline to each module from all their inputs and modules can manipulate them in several ways on the way through (renaming, mapping values). It is then possible for a module further down the pipeline to use information about the provenance of a module (and, thanks to the feature above, a datatype).