Big bugfix and several small enhancements
Big bug: output names in input specifications got lost during alternative module expansion.

Also, keep track of the modules that got expanded while the pipeline config was read to facilitate the alias module.

New alias module to allow a particular output from a module to be aliased as if it were the default ouput for a new module instance.

Some small enhancements to the split module.