Continue to later modules if one fails
A major boon for anyone running large pipelines that aren't entirely linear. If one module fails, we don't drop out completely, but continue to the next module that can be run. Any modules that depended on the one that had an error must, of course, be skipped, but if your pipeline has branches there will often be others that can be executed.

This can avoid a lot of wasted time when your long-running pipelines are sitting not being run until you come back after your holiday to discover that a small error early on prevented the rest from being run.

Added an option to the run command to go back to the old behaviour, for those who don't like efficiency. Also added some further error output to make it easier to trace what went wrong when examining your sorrowfull terminal output post-holiday.