Don't trust the output from failed modules
If a module fails, when trying to execute other modules in the pipeline, don't just check whether it looks like their data is ready. It might be a valid dataset, but if the module building it failed, it's presumably not complete, so assume it's not valid, and don't execute those later modules.