Use threaded pool impl when running with 1 proc
Previously used multiprocessing with multiprocessing document map modules (created with the factory), even if only one process was requested.
This can cause difficulties on some systems with some software, so safest to use the threaded implementation and not spawn any new processes if only one process has been requested, even though it makes the code a little more convoluted (which is why I didn't implement this from the beginning).