Changed behaviour of force-rerun option
Used to start by resetting the execution state of the module, but this meant that any existing data would be lost, including useful caches or temporary data. Removed this.

This could mean that forced reruns behave oddly in some circumstances, but there's always the option of resetting execution manually, which is the safe way to ensure it starts from scratch.