Enhanced handling of modvars
Now modvars coming from multiple-inputs get combined into a list, instead of being conjoined into a string, meaning that we can use the multiple values explicitly later. If such a modvar is substituted, most of the time an error is produced, since it doesn't make sense to substitute in a list. However, if you try to substitute it into an input specification, it is expanded into a list of input specs, using each of the values. This allows it to be used in a multiple-input, taking inputs corresponding to each of the values in the modvar list.