Changed behaviour of map filters
Previously, when you ran a document map module as a filter, during postprocessing, before the result was yielded, the map module's writer was always used to convert the output document to raw data and then the reader to convert it back to the programmatic data structure. This easily gets very inefficient and was only there to add some flexibility to the reader/writer-writing -- they didn't necessarily have to be symmetric. But in practice, that flexibility is rarely used, and by removing this we make the filters generally a lot faster, especially if the module itself doesn't do very much processing.