Added super-detailed debugging
When debugging a pipeline, it can be useful to step through what it's doing interactively to get to the bottom of how the output looks at different points in the pipeline. This is particularly useful when chaining together a whole load of modules as filters.

Added a new switch --step, which enables a special debugging that wraps a load of functions to output lots of information. It's also interactive, so you can step through slowly and request that certain types of output get skipped in future when they come up.