Added email notification feature
Pimlico can now send you emails to update you when a module fails or give you a final report when execution finishes.

Added email sending utilities, a cmd-line command to test the email configuration in the current environment and a new switch for the run command to enable email report sending.