Documentation of multistage modules
Wrote a guide to defining and using multistage modules, something which has had very scanty documentation up to now.

Also added a tiny feature that allows you to run all stages at once, as I realised while writing the docs that this is a useful thing (and now very easy to add -- it wasn't when I implemented multistage modules).