Added option to run all runnable modules
Also changed the way multistage modules get added to the pipeline. This was causing a problem, as the module_order list contained both the individual stages and the main module, which doesn't really make sense. Now the special case for multistage modules is moved elsewhere (which also makes it easier to follow what's going on) and the main module is made available by name, but not include in the order.