New module to concatenate corpora
Very useful particularly for handling input data, where multiple corpora may need to be created to allow different input processing, but we then want to merge them.

Simply concatenates the list of files, or of archives in the case of a tarred corpus.

Works for tarred corpora, or falls back to iteratable corpora if not all the inputs are tarred corpora.