Proper dynamic subclassing of tarred corpora
Instead of instantiating a tarred corpus (or iterable corpus) and setting the datapoint_type, we should dynamically subclass the base type with a new datapoint_type. This is because class methods access the datapoint type and will not have access to the overridden type is it's set on the instance.

This need for a lot of dynamic creation of types is one of the key things I intend to get away from with the new datatype system that's on its way sometime soon. For now, we have to stick to doing it this way.