Small changes to COPY directive in config files
Can now be applied multiple times, copying from a module that itself uses a COPY directive. It's then important that these are processed in the order they're specified in the file.

Now allow copying of all parameters, including module type. These can all be overridden, so there's no reason not to.