Added new module test framework
New script and routines for running small test pipelines, allowing us to build test suites for builtin modules.

This is an alternative to unit tests (though it would be appropriate to add a number of those too), more appropriate for testing the functionality of modules.

This is to address #7, preparing the way for the new datatype functionality.