New input reader module factory
This is part of the ultimate aim to phase out the use of input_module_options on datatypes altogether and replace it with input reader modules. This factory makes it easy to create such modules for reading iterable corpora.

Added a first example of its use -- the raw text files reader. This is to replace the RawTextFiles datatype (the main use of UnnamedFileCollection): added documentation to explain this.

Changed module tests to use the new input reader.