Changed the way input reader factory works again
I'm working this out as I go along, but no-one but me is using this yet, so no worries. This way of defining an input reader makes it easier to make processing code extensible. In particular, raw files reader now allows you to easily add processing when overriding to create a reader that does something more.