Added input reader for Facebook's embedding format
Reads into Pimlico's (Gensim-based) internal data structure, so that they can be used in pipelines.

Modified BaseModuleInfo to add a new type of data availability check. This is on modules, not their input datatypes, so that input modules can specify extra data checks on the basis of their options, where no input datatype is available to perform those checks.