Split command should copy input metadata to output
Fundamentally, the problem is that we write the output corpus in a generic way, not using the specific datatype's writer. Assuming that all that is specific to the datatype at the corpus level is stored in the metadata (which I think should usually, if not always, be the case), if we copy over the input corpus' metadata to the output, allowing things like length to get overridden, we should end up with a useable corpus.