Added possibility to expand alternatives in list
Previously, an input parameter that contained a comma-separated list of inputs (multiple inputs) never expanded any of the input module specifications into multiple modules, as is done with single inputs. This was because it's not clear how they should be expanded: when multiple input specs in the list have different expansions, what is the correct behaviour?

Now list inputs do get expanded. If everything in the list has the same number of alternatives, they get tied together (like with Python's zip) and the result is an alternative for each expanded input. If some can be expanded and others not, the non-expanding ones get repeated for each of the expansions. If there are different numbers of expansions for different items in the list, an exception is raised.

Note that this is only for expansion that propagate through the pipeline, i.e. where an unexpanded module name is used to refer to all of the expanded (alternative) versions of the module. Where alternatives are given explicitly using a |, these are split before the list items, i.e. it's considered to be multiple alternative lists, not alternatives within the list.