Added valid docs count to inputs
This is useful for a lot of things, but note that this feature is not complete. The count cannot be relied upon, except when it's come straight from the input reader. Further down the pipeline, there might be more invalid documents, but the count is not currently updated. All internal modules should update (or at least remove) the count, just as they do currently with the length count.