Large batch of work on #1
First basic implementation of new datatypes system. The basic structure for datatypes is done here, but I've not yet updated any of the core code that uses them (module loading, etc).

This also includes several of the core datatypes, converted to the new system, including iterable corpora, with the new way of specifying document types.