Getting subclassing working
Added unittests for some datatypes and in the process got the subclassing of readers, writers and reader setups working correctly.

Next job will be to add lots more datatypes and it's now easy to add test cases for each one.

Added a list of the datatypes that need to be converted still.