Working towards getting pipelines running
Not converted all datatypes yet and not updated many built-in modules, but the next step is to get pipelines running, so we can start using the half-finished, broken new version of the codebase!

Working through the test pipelines. Once they're all running OK, we've got a long way towards having the new datatypes system functional.