Changed how documents are instantiated
Interface for instantiating documents from data point types is now much nicer, easier to read, more intuitive. We now use kwargs instead of a dictionary with the from_internal flag to create a doc from internal data, which, after all, is the more common way to need to do it.

This required that data point types know what keys the document's internal data should have. This is good anyway, as it makes the data structure easier to understand and means we can include this information in the docs (when we start generating datatype docs).