Extended browser functionality
Browser is no longer only for iterable corpora, but old iterable corpus browser is now a special case.

All datatypes can implement a browser which may be any kind of interative (or even non-interactive) browser to display the data. For iterable corpus, this is now just the old browser.

Added one other kind of browser, which is a simple file selector and displayer for named file collections. Lots of datatypes are subtypes of this, so it provides some basic way of viewing many datatypes.