Marked datatypes branch as 1.0 rc
Since no-one is using master in parallel to the datatypes update and I'm not really maintaining it separately, now that the datatypes branch is quite mature, I'm going to merge it back into master as a RC for v1.0. I've marked the current master as release v0.9 and tagged, so that will be the last available version before the datatypes update is applied in release 1.0.