Changed naming of automatic subclasses
Datatype nested classes are subclassed automatically. The automatic class names meant that the classes didn't show up at all in the documentation. Now they have the simple names "Reader", "Writer" and "Setup", which are the names that are used to access them, so Sphinx includes them in the documentation.

There's still a problem with referring to parent classes, but it may be we can't fix that.