Changed structure of input reader factory/classes
Moved a lot of what was in the factory function into base classes, allowing them to be used directly without the factory if desired. The factory is still the easiest way to create an input module, but we now have another alternative that is more flexible w.r.t. subclassing and overriding parts of the behaviour.

The factory still provides the same interface and the same behaviour as before.