CorpusGroupReader now inherits from IterableCorpus
Some of the functionality provided by IterableCorpus' reader was being lost in CorpusGroupReader, since it didn't subclass it.

Added a new method of IterableCorpus' reader to allow init operations to take place before the datapoint type's init method is called. Moved all of CorpusGroupReader's init into there, since it has to do this stuff before the metadata can be read.