Added allow_skip_output option to map modules
It's now possible to have map modules that sometimes return None for one or more of their outputs (from the process_document function) and just don't write anything for that output.

This was necessary to have a module that splits the data into multiple corpora. Previously, you couldn't use the map module factories for this case, but this simple extension means you can. The default behaviour is the same, though: all outputs are expected for every document.