Allow FAILED map modules to resume
Previously only a status of PARTIALLY_PROCESSED would cause the module to resume where it left off, which only really covers where the user exits in the middle of processing. If there's an error, the module is left with FAILED status. Now, if the progress has been appropriately stored, the module will pick up where it failed in this case.