Updated filter mode map execution for new dtypes
Finally got round to updating the filter mode execution of document map modules.

Abstracted the core map execution routine from the main map executor, so that we can maximally share the code for that part of the process.

The filter mode wrapper now replaces output corpus readers instead of dynamically providing a new datatype, as we needed to do in the old datatypes system. The datatypes can now be treated separately, and are identical to those used outside filter mode.