Fixed buffering problem with doc map processes
Important fix for how results are buffered when they come in from worker processes in a document map module (processes or threads, depending on implementation). The problem arose when a module's doc processing was very fast, so the bottleneck was writing the results out to disk. The queue buffering the output from the workers got bigger and bigger, eventually resulting in a memory problem if the dataset was very large.

Now I've simply set a (quite large) limit on the output queue size, so buffering is limited to ~10 docs per worker, if it's taking a long time to postprocess them.

This fixes the memory problem, but also, I think, makes execution faster. Another effect is that the time estimate is more stable and accurate.