Made list_archive_iter more incremental
Change the way the members of tar archives are read in grouped corpus when only reading filenames. Now we don't try to read a full archive at a time, but one file at a time, meaning that we can get as far as the last document on archives that are partly written and have errors at the end. (Needed for recover command)