Improved counting of existing docs when appending
When a doc map module appends a partially written corpus, it counts up the existing docs. This doesn't always seem to be working. I'm not certain yet whether this fix solves that, as I'm not sure yet under what circumstances it's happening, but it's a more reliable way to count the docs anyway.