Changed format of raw_data
Used to be documented as being a unicode, but actually was a string storing raw data (utf-8 encoded in the case of text).

The correct way to make that Py3-compatible, then, is to use a bytes() to store the raw data. Changed to this and updated docs to reflect it. There will be more datatypes that need to be corrected for this, but the most common ones are covered here.