Removed untokenize module
This was only added as a workaround to a problem caused by the old datatypes system.

Now the new datatypes system is working, there's no need for this any more.

Note that there could be a use for a more sophisticated detokenization module, but that's not what this was anyway.