Changed scipy and numpy deps to normal pip install
For some reason, Scipy and Numpy dependencies were marked as ones where you had to install them system-wide. I'm not sure why this was: presumably something to do with the history of Scipy and Numpy development.

Nowadays, as far as I know, there's no reason not to install Scipy and Numpy in a virtual environment, so I've changed the standard dependencies to normal pip installations.