Fixed merge conflict.
Everything seems to be working nicely in both Python 2 and 3. We have a load of tests, which are passing in both versions.

There are still some modules to be updated to the new datatypes system and this involves checking/fixing them for Python 3. However, since these all need to be gone through anyway before the next major release.