Allow non-document output from process_document()s
A process_document() method in document map modules can now output a bytes object (raw data for doc), a dict (internal data for doc), or a Document instance. The latter used to be the only option, but resulted in unnecessary pickling and unpickling. In future, one of the first two will be the recommended output type, since it should result in much faster process communication.

This change is not yet properly documented. In future, it should be well document in docs about writing doc map modules.