Added new Pimarc format
Most of basic implementation of file reading and writing now complete but not tested.

This will be used to replace tar as the archive format for storing grouped iterable corpora. For now, it's not used anywhere, but once it's been tested, it will become the standard format for writing GroupedCorpora. We will still support reading from old tar-based corpora, but will detect which format has been used.