Added new shuffle module and moved old one
The new shuffle module is better, as it uses the random access to the input corpus allowed by the Pimarc storage. However, it can only be used where the input is stored, not dynamically produced. So the old implementation is still needed, now moved to shuffle_linear.

Updated the shuffle_linear implementation to use Pimarc storage for the temporary bins used during shuffling. This should speed things up and makes the implementation much cleaner.