Fix metadata writing on error
Metadata should be written out when exiting the context manager for a
PimlicoDatatype regardless of whether there was an error or not. This is
important, as it can contain information about partially executed state,
as in the case of a corpus that's being written out, where it stores how
far the module got.