Removed trust_length option from grouped corpus
We never use this anywhere and I don't think there's any reason why we
would.

Now always checks the length of a corpus before appending to it, to be
sure to avoid problems with out-of-date metadata.

This used to be very time consuming, but with the Pimarc storage format,
it's now generally very quick.