Fixed problem shutting down threaded doc mapper
Synchronization problem with queues. The input reader queue is getting
shut down while the processor is waiting for a document, which was
causing an exception to be raised in the worker process while shutting
down.

This is not a problem: we just have to catch the exception and ignore
it. The stopped flag is set by this point, so we let the process exit
naturally when this happens.