Fixed references to base classes in docs
Base classes of dynamically created Reader, Writer, Reader.Setup and
Document classes were not correctly referring to the actual base class
in the docs. Fixed references by correctly setting __qualname__ and
__module__ on the generated classes.

Also added some tricks for the docstrings, so we only use class docstrings
that are set explicitly, not ones that get inherited.