Improved error output on PY3
Chaining of errors using "raise from" via future's wrapper was losing a
lot of debugging information, including traceback for the cause of the
error. Instead, we now use the real "raise from" on PY3 and only use
future's replication of it on PY2. This means that there'll be less good
debugging info on PY2, but that's not a priority anyway.