Added batch processing possibility to input feeder
Doc map processes can accept multiple documents at a time, if the input
feeder gives them multiple.

The input feeder batches the docs before sending them to the workers.

Batch size is controlled by a parameter on the input feeder, currently
always set to 1.

This will be changed in future.