Doc map progressbar smoother
Previously, progress bar was based on docs yielded from the mapper. If
one process gets stuck for a while, results from the others get
buffered. Previously, the progress bar would then sit still, even though
results were coming in.

Now the progressbar is updated within the doc mapper, based on docs
yielded by the worker processes, before buffering is done to get them in
the right order. This results in a smoother and more realistic
representation of the progress.