Relicensed and updated copyright
Re-licensed Pimlico to move from the GPL to LGPL. I should have chosen
this from the beginning, since it allows Pimlico to be used in
commercial, non-OS systems under restrictive licenses.

Note that I am currently the sole contributor to the codebase, so own
the copyright for all code, making this an OK thing to do.

Updated the copyright messages throughout the codebased, including the
new license and this year's copyright.