Make it easier to see last error details
If an error is encountered during execution when not in debug mode, the
full error output is sent to a file. Instead of copying and pasting the
file path to see it, you can now just re-run the previous run command
with the '--last-error' or just '-e' switch to view the error.