Major speedup in grouped corpus writing
Removing flushing after each new document addition. This was to avoid
output corpora getting out of date if a process dies, so they don't get
left in an inconsistent state and so we get the latest data when
restarting. However, not doing this speeds up the writing (and thereby
the whole processing for many modules) that it's not worth it!