Benchmarking and efficiency for doc maps
Added new benchmarking tool for doc map module executor to investigate
what is taking most time and holding up the processing.

Updated example pipelines to contain one that takes longer to execute,
for investigation.

Investigation showed that the bottleneck was writing the metadata after
receiving each document's result. This was because we were flushing the
metadata file after writing to ensure we had exact progress information
if the process stopped partway. This is no longer important, since the
newer Pimarc storage system. The flushing has already been removed from
the archive writing and the actual written documents are checked when
restarting anyway.

Removing this flushing brings a huge improvement in speed when the
individual document processing is fast.